@@ -9,6 +9,8 @@ import java.util.Date;
 import static ru.javawebinar.topjava.model.AbstractBaseEntity.START_SEQ;
 
 public class UserTestData {
+    public static TestMatcher<User> USER_MATCHER = TestMatcher.usingFieldsComparator("registered", "roles");
+
     public static final int USER_ID = START_SEQ;
     public static final int ADMIN_ID = START_SEQ + 1;
 
@@ -25,6 +27,4 @@ public class UserTestData {
         updated.setCaloriesPerDay(330);
         return updated;
     }
-
-    public static TestMatcher<User> USER_MATCHER = TestMatcher.of("registered", "roles");
 }
